backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
starlette==0.36.3
pydantic==2.7.1
websockets==12.0

